Youtube-Shorts-Generator/requirements.txt
asgiref
Django
Pillow
requests
social-auth-app-django
python3-openid
certifi
charset-normalizer
cryptography
defusedxml
idna
oauthlib
PyJWT
requests-oauthlib
wheel
setuptools
termcolor
schedule
prettytable
webdriver_manager
selenium_firefox
selenium
g4f[all]
moviepy
Pillow==9.5.0
yagmail
assemblyai
srt_equalizer
undetected_chromedriver
platformdirs
google-generativeai
gtts
Brotli
edge-tts
playsound
telethon
PyExecJS
psutil
#TTS
